refactor(test): extract shared test helpers and fix test infra
Move testdataFile, createCmd, runCmd, and deleteKustomizeDeployment to
helpers.go so both e2e and quota_e2e build tags can use them. Add
QUOTA_HELPER_IMAGE kind-load to SetupSuite, fix time.Tick to
time.NewTicker, add commonLabels to kustomization files, and rename
storageclass.yaml to storage-class.yaml.
