+++ test/testdata/custom-path-pattern/kustomization.yaml
@@ -6,6 +6,8 @@ resources:
 - pod.yaml
 - pvc.yaml
 images:
+- name: local-path-provisioner
+  newTag: dev
 - name: rancher/local-path-provisioner
   newTag: dev
 labels:
@@ -14,3 +16,5 @@ labels:
     app: local-path-provisioner
 patches:
 - path: local-path-config.yaml
+commonLabels:
+  app: local-path-provisioner
