+++ test/testdata/security-basic-traversal-path/kustomization.yaml
@@ -1,6 +1,8 @@
 apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
 resources:
-- storageclass.yaml
+- storage-class.yaml
 - pvc.yaml
 - pod.yaml
+commonLabels:
+  app: local-path-provisioner
