+++ test/testdata/security-basic-traversal-path/storage-class.yaml
