+++ test/testdata/skip-path-pattern-check/kustomization.yaml
@@ -14,3 +14,5 @@ labels:
     app: local-path-provisioner
 patches:
 - path: local-path-config.yaml
+commonLabels:
+  app: local-path-provisioner
